Try the auth options
This is mostly a failure, but some of the Obj-C code is not
trivial to write, so I'll have it in this commit, and then
remove the failed attempt in the next one.

Findings:
- `kSecUseAuthenticationUI` does not work for me at all.
  "skip" will not skip anything, but prompt, "fail" will
  not fail, but "prompt".
- `kSecUseAuthenticationContext` does not work either.
  If I set it to an initialized LAContext, I still get
  a separate password prompt for each item.
  If I set a credential to it with `setCredential()` and
  `LACredentialTypeApplicationPassword` I still get password
  prompts.
  If I authenticate with this context, that does not seem
  to work either.

So we'll remove the wrapper to `kSecUseAuthenticationUI`
completely, since it does not work, and document that
if `return_data` is set, then you might get a bunch of
password prompts.